Move directory to a less irritating name.
