+++ opencog/atoms/CMakeLists.txt
@@ -1,7 +1,7 @@
 
-ADD_SUBDIRECTORY (bind)
 ADD_SUBDIRECTORY (core)
 ADD_SUBDIRECTORY (execution)
+ADD_SUBDIRECTORY (pattern)
 ADD_SUBDIRECTORY (reduct)
 
 INSTALL (FILES
