+++ opencog/atoms/pattern/BindLink.cc
