+++ opencog/atoms/pattern/PatternLink.cc
