+++ opencog/atoms/pattern/PatternLink.h
