+++ opencog/atoms/pattern/PatternUtils.cc
