+++ opencog/atoms/pattern/PatternUtils.h
