+++ opencog/atoms/pattern/README.md
