+++ opencog/atomspace/AtomSpace.cc
@@ -33,7 +33,6 @@
 #include <opencog/atomspace/ClassServer.h>
 #include <opencog/atomspace/Link.h>
 #include <opencog/atomspace/Node.h>
-// #include <opencog/atoms/bind/DeleteLink.h>
 #include <opencog/atomspace/types.h>
 #include <opencog/util/Logger.h>
 #include <opencog/util/oc_assert.h>
