+++ opencog/query/FuzzyMatch/FuzzyPatternMatch.cc
@@ -22,7 +22,7 @@
  */
 
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atoms/bind/PatternLink.h>
+#include <opencog/atoms/pattern/PatternLink.h>
 
 #include "FuzzyPatternMatchCB.h"
 #include "FuzzyPatternMatch.h"
