+++ opencog/query/Implicator.cc
@@ -23,7 +23,7 @@
 
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/SimpleTruthValue.h>
-#include <opencog/atoms/bind/BindLink.h>
+#include <opencog/atoms/pattern/BindLink.h>
 
 #include "BindLinkAPI.h"
 #include "DefaultImplicator.h"
