+++ opencog/query/PatternMatch.cc
@@ -23,8 +23,8 @@
 
 #include <opencog/util/Logger.h>
 
-#include <opencog/atoms/bind/BindLink.h>
-#include <opencog/atoms/bind/PatternLink.h>
+#include <opencog/atoms/pattern/BindLink.h>
+#include <opencog/atoms/pattern/PatternLink.h>
 
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/ClassServer.h>
