+++ opencog/query/PatternMatchEngine.cc
@@ -24,7 +24,7 @@
 #include <opencog/util/oc_assert.h>
 #include <opencog/util/Logger.h>
 #include <opencog/atomutils/FindUtils.h>
-#include <opencog/atoms/bind/PatternUtils.h>
+#include <opencog/atoms/pattern/PatternUtils.h>
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/Link.h>
 #include <opencog/atomspace/Node.h>
