+++ opencog/rule-engine/backwardchainer/BackwardChainer.cc
@@ -32,7 +32,7 @@
 #include <opencog/atomutils/FindUtils.h>
 #include <opencog/atomutils/Substitutor.h>
 #include <opencog/atomutils/AtomUtils.h>
-#include <opencog/atoms/bind/PatternLink.h>
+#include <opencog/atoms/pattern/PatternLink.h>
 
 using namespace opencog;
