+++ tests/query/PatternCrashUTest.cxxtest
@@ -23,7 +23,7 @@
 #include <opencog/util/Config.h>
 #include <opencog/util/Logger.h>
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/atoms/bind/PatternLink.h>
+#include <opencog/atoms/pattern/PatternLink.h>
 #include <opencog/atomutils/FindUtils.h>
 #include <opencog/guile/SchemeEval.h>
 #include <opencog/guile/load-file.h>
