+++ tests/query/PatternUTest.cxxtest
@@ -25,7 +25,7 @@
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/query/DefaultPatternMatchCB.h>
 #include <opencog/query/InitiateSearchCB.h>
-#include <opencog/atoms/bind/PatternLink.h>
+#include <opencog/atoms/pattern/PatternLink.h>
 #include <opencog/util/Config.h>
 #include <opencog/util/Logger.h>
